Added Documentation Added support for SDK 21 Added ScreenManager class to handle hiding/showing panels
