+++ android/controller/src/main/res/drawable-v23/sliders_background.xml
@@ -9,6 +9,22 @@
   -->
 
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:width="24dp"
+        android:height="24dp"
+        android:drawable="@drawable/left_notch"
+        android:gravity="center_vertical"
+        >
+    </item>
+
+    <item
+        android:width="24dp"
+        android:height="24dp"
+        android:gravity="right|center_vertical">
+        <rotate android:fromDegrees="180.0" android:toDegrees="0.0"
+            android:drawable="@drawable/left_notch"/>
+    </item>
+
     <item>
         <shape android:shape="rectangle">
             <stroke
